fix issue 2018 for redis error when started not used
