+++ tunnel-server/src/main/java/com/alibaba/arthas/tunnel/server/app/ArthasTunnelApplication.java
@@ -2,11 +2,10 @@ package com.alibaba.arthas.tunnel.server.app;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.cache.annotation.EnableCaching;
+import org.springframework.boot.autoconfigure.data.redis.RedisAutoConfiguration;
 
 @SpringBootApplication(scanBasePackages = { "com.alibaba.arthas.tunnel.server.app",
-        "com.alibaba.arthas.tunnel.server.endpoint" })
-@EnableCaching
+        "com.alibaba.arthas.tunnel.server.endpoint" },exclude = {RedisAutoConfiguration.class})
 public class ArthasTunnelApplication {
 
     public static void main(String[] args) {
