+++ tunnel-server/src/main/java/com/alibaba/arthas/tunnel/server/app/configuration/EmbeddedRedisConfiguration.java
@@ -3,11 +3,13 @@ package com.alibaba.arthas.tunnel.server.app.configuration;
 import org.springframework.boot.autoconfigure.AutoConfigureBefore;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
+import org.springframework.boot.autoconfigure.data.redis.RedisAutoConfiguration;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 
 import com.alibaba.arthas.tunnel.server.app.configuration.ArthasProperties.EmbeddedRedis;
 
+import org.springframework.context.annotation.Import;
 import redis.embedded.RedisServer;
 import redis.embedded.RedisServerBuilder;
 
@@ -16,13 +18,14 @@ import redis.embedded.RedisServerBuilder;
  * @author hengyunabc 2020-11-03
  *
  */
+@AutoConfigureBefore(value = {TunnelClusterRedisStoreConfiguration.class, TunnelServerConfiguration.class})
 @Configuration
-@AutoConfigureBefore(TunnelClusterStoreConfiguration.class)
+@ConditionalOnProperty(name = "arthas.embedded-redis.enabled", havingValue = "true")
+@Import(RedisAutoConfiguration.class)
 public class EmbeddedRedisConfiguration {
 
     @Bean(initMethod = "start", destroyMethod = "stop")
     @ConditionalOnMissingBean
-    @ConditionalOnProperty(prefix = "arthas", name = { "embedded-redis.enabled" })
     public RedisServer embeddedRedisServer(ArthasProperties arthasProperties) {
         EmbeddedRedis embeddedRedis = arthasProperties.getEmbeddedRedis();
